ensure we stop processing program after processing stop_diagram message
If the user ran a program with no data storage block, we would continue to process the program (and affect the relay) after handling the stop_diagram message.  This was because the comparion checking if stop_location == recording_location failed (a unique case when we have no recorded data) and we bailed before setting self-diagram = none.  By leaving the diagram with the previous state (instead of none), we continued to process the program in the main while loop.  Be sure to set diagram to none to prevent this from happening.
[#160222226]